fix (Modules) Run grunt for module root through steroids-cli
Previously, grunt compilation in root was done through the mobile harness' Gruntfile.
Due to an apparent issue in Node 0.12, this would cause the grunt task to be left
waiting for input. Because no input would be given to the task when running through
steroids connect, this would prevent steroids connect from continuing because the
grunt task runner would never complete.